backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.5.2
Flask-Cors==4.0.0
Flask-Migrate==4.0.4
python-dotenv==1.0.0
psycopg2-binary==2.9.7
bcrypt==4.0.1
Marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
Werkzeug==2.3.7
gunicorn==21.2.0
requests==2.31.0
python-dateutil==2.8.2
